+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Only send SNS message when the Transit Gateway attachment is in pending approval state.
+- Add cluster name as tag `Name` when creating the Transit Gateway attachment.
 
 ## [1.3.0] - 2023-01-17